coloquei uma informação do final da questão onde tem o desconto

+++ feitos-por-mim/exercicio-casa/atividade_27.py
@@ -1,7 +1,9 @@
+
+import math
 #pedir a quantidade de morangos e maças, separadas
 
-quant_morangos = int(input("Insira aqui o quilo de morango que deseja: "))
-quant_macas = int(input("Insira aqui o quilo de maçãs que deseja: "))
+quant_morangos = float(input("Insira aqui o quilo de morango que deseja: "))
+quant_macas = float(input("Insira aqui o quilo de maçãs que deseja: "))
 
 #definir o valor para 5kg e acima de 5kg, cada. 
 
@@ -15,6 +17,8 @@ if quant_morangos <= 5:
 else: 
     total_valor_morangos = quant_morangos * valor_acima_5kg_morangos
 
+
+
 if quant_macas <= 5: 
     total_valor_macas = quant_macas * valor_5kg_macas
 else: 
@@ -27,8 +31,13 @@ total_sem_desconto = total_valor_macas + total_valor_morangos
 if (quant_macas + quant_morangos) >= 8 or total_sem_desconto >= 25.00:
     aplica_desconto = 0.10 
     total_valor_com_desconto = total_sem_desconto * (1- aplica_desconto)
+                                                    #Podemos colocar 0.90, já que queremos 10% de desconto, se fosse 20% de desconto seria 0.80 e assim por diante, (valor total - o desconto.(ta confuso mas faz sentido na minha cabeça kk))
+
+
 else: 
     total_valor_com_desconto = total_sem_desconto
+
+
 
 print(f"O valor total dá igual a R${total_valor_com_desconto:.2f}")
